llm/google/gemini/requirements.txt
python-dotenv
langchain
langchain-google-vertexai
langchain-google-genai
google-generativeai
dashscope
qdrant-client
faiss-cpu
sentence-transformers
